Add new way to list files for loading
The methods used in the `LoadDSL` for listing files can now take a
string and an array of strings. The first string will be treated as a
filename, and the array will be treated as a list of directories in
which the file can be. `cascade 'foo.yml', ['.', '..']` is equivalent to
`cascade './foo.yml', '../foo.yml'`.